requirements.txt
aioredis==2.0.1
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
autocommand==2.2.2
Automat==22.10.0
Babel==2.14.0
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
cheroot==10.0.1
CherryPy==18.9.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
defusedxml==0.7.1
dictor==0.1.11
distlib==0.3.8
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
django-guardian==2.4.0
django-log-request-id==2.1.0
django-reversion==5.0.11
django-rosetta==0.10.0
django-simple-history==3.5.0
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-flex-fields==1.0.2
drf-yasg==1.21.7
elementpath==4.1.5
filelock==3.13.1
flake8==7.0.0
flower==2.0.1
gevent==24.2.1
greenlet==3.0.3
hiredis==2.3.2
humanize==4.9.0
hyperlink==21.0.0
identify==2.5.33
idna==3.7
importlib_metadata==7.1.0
incremental==22.10.0
inflect==7.2.1
inflection==0.5.1
install==1.3.5
jaraco.classes==3.4.0
jaraco.collections==5.0.1
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.0
kombu==5.3.7
mccabe==0.7.0
more-itertools==10.2.0
msgpack==1.0.8
msgpack-python==0.5.6
nodeenv==1.8.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
platformdirs==4.1.0
polib==1.2.0
portend==3.2.0
pre-commit==3.6.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pyflakes==3.2.0
PyHamcrest==2.1.0
PyJWT==1.7.1
pyOpenSSL==24.1.0
pyparsing==3.1.2
pysaml2==7.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.31.0
service-identity==24.1.0
setuptools==68.2.2
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tempora==5.5.1
tornado==6.4
Twisted==24.3.0
txaio==23.1.1
typeguard==4.2.1
typing_extensions==4.11.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.13
wheel==0.41.2
xmlschema==3.0.1
xmltodict==0.13.0
zc.lockfile==3.0.post1
zipp==3.18.1
zope.event==5.0
zope.interface==6.3
